pyproject.toml
[tool.poetry]
name = "yfai"
version = "0.1.0"
description = "本地对话式控制台 - 统一接入百炼与Ollama，MCP Client集成与本地控制能力"
authors = ["YFAI Team"]
readme = "README.md"
packages = [{include = "yfai"}]

[tool.poetry.dependencies]
python = "^3.11"
PyQt6 = "^6.6.0"
PyQt6-WebEngine = "^6.6.0"
qasync = "^0.27.0"
pydantic = "^2.5.0"
httpx = "^0.25.0"
tenacity = "^8.2.3"
dashscope = "^1.14.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
cryptography = "^41.0.7"
psutil = "^5.9.6"
watchdog = "^3.0.0"
SQLAlchemy = "^2.0.23"
structlog = "^24.1.0"
typer = "^0.9.0"
numpy = "^1.26.2"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
pywinauto = {version = "^0.6.8", markers = "sys_platform == 'win32'"}
pynput = "^1.7.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-qt = "^4.2.0"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.6"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yfai = "yfai.main:main"

